Add --dialect option, remove dialect sniffing
Python's csv.Sniffer really looks like it's going to be useful, and
then it occasionally screws up royally. And it *thinks* it's done a
bang-up job, so you don't get any indication of failure.

I think the only thing for it is to throw csv.Sniffer in the garbage. Instead of having a million options for delimiters and quoting, we'll
just offer the standard 'excel' and 'excel-tab' dialects. Write your
files according to those specs.

This does mean your scoresheet and data need to be in the same format.
Sorry.