Add imputation functions
Added mean_imputed(), sum_imputed(), and imputed_fraction() functions.
Each of these replaces non-numeric values in their measure with
the mean of the numeric values — so [1, '', 3, 5] has a mean_imputed
of 3 and sum_imputed of 12.

imputed_fraction() returns the fraction of the values that would be 
imputed in those functions, so with the same example, the
imputed_fraction() is 0.25.

Also test and document these.