Refactor score_data.py
Previously, score_data was basically one gigantic main() method that did
everything from loading data to scoring it and printing it out. This
meant that when I wanted to use it from another script, I was capturing
stdout, which... I mean it works, but yuck.

This splits score_data up into more modular parts. Most importantly,
scoring and printing data are separate functions, so you can score
without printing, which we'll use in score_multi.

This also removes the --ignore-missing option, which appears to have
been completely aspirational anyhow. (Being able to filter data is still
arguably a good idea, but that's for a later project and more thought.)