Add score_multi
This takes a CSV file that defines a set of scoring operations -- input,
scoresheet, and output, and runs score_data on each row. Input,
scoresheet, and output are each formatted, with format_map() for each
row of the definition scoresheet. (In this case, the mapper is a special
one that'll return the {key} if that's missing from the row, instead of
crashing.)

The use of format_map means you _can_ do things like padding filenames
to particular lengths, if you want to.

Anyhow, this should be fast and work and be good. Good-ish.

Closes #32.